A beginning of a wrapper around the node-github
Having a wrapper will ease the test-driving of the connector code, and
will allow me easily change node-github library with something else when
needed.

I’m not useing mocks here because this is boundary code and I want to
make sure it works. Its tests will be disabled most of the time and will
be run only on library updates, API changes, or something like that, so
there is no reason to worry about their speed.

I am putting the token here because I don’t think the risk is negligible
compared to the effort of setting up a configuration setting.