GitHubSession.getOrganisations
Moved from GitHubSession.prototype.getOwnRepos to this.getOwnRepos in
the class definition because the latter is more concise. Also, it lets
me keep the GitHubApi instance private and makes for less boiler place
like `var github = this`.

Extracted the errorHandler helper function for the catch case of the
promises in tests.